dashboard/requirements.txt
flask==3.0.3
boto3==1.34.69
python-dotenv==1.0.1
